Move measure class to separate module (#180)
Move FIXME(#142)
